refactor: move MemoryDB struct to test utils
