+++ pkg/trie/triedb/util_test.go
@@ -1,7 +1,4 @@
-// Copyright 2024 ChainSafe Systems (ON)
-// SPDX-License-Identifier: LGPL-3.0-only
-
-package db
+package triedb
 
 import (
 	"bytes"
